+++ dotnet-roslyn/analyzers/Audacia.CodeAnalysis.Analyzers/CHANGELOG.md
@@ -5,6 +5,9 @@
 - Added new rule "AssertionScopeForMultipleAssertions" (ACL1019).
     - This rule checks that if there are more than 2 assertions within a test method then they are wrapped in an AssertionScope. This provides better failure messages and allows all assertions to be evaluated even if one fails.
     - This rule works with Xunit, Shouldly, and FluentAssertions syntax.
+- Added new rule "AssertionRequired" (ACL1020).
+    - This rule checks that all Xunit test methods have at least one assertion, ensuring that tests are meaningful and validate expected outcomes.
+    - This rule works with Xunit, Shouldly, and FluentAssertions syntax.
 
 ### Changed
 - No changes to existing functionality